requirements.txt
asgiref==3.2.10
cachetools==4.2.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-heroku==0.3.1
django-utils-six==2.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
